leddar_one: split into separate header, implementation, and main
 - remove default serial port
